[random] Fixes and improvements
